Refactored squad generation from XWS into its own class
